pyproject.toml
[project]
name = "istore"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.1.2",
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-pymongo>=3.0.1",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.15.0",
]

# pyproject.toml
[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F", "I"] # "I" is for Isort (Import sorting)
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
